refactor: organize components by features folders
