+++ src/components/Configs/Insights/ConfigAnalysisLink/ConfigAnalysisLink.tsx
@@ -1,10 +1,10 @@
 import { useState } from "react";
-import { ViewType } from "../../../types";
-import ConfigInsightsIcon from "../Insights/ConfigInsightsIcon";
-import { DescriptionCard } from "../../DescriptionCard";
-import { ConfigIcon } from "../../Icon/ConfigIcon";
+import { ViewType } from "../../../../types";
+import ConfigInsightsIcon from "../ConfigInsightsIcon";
+import { DescriptionCard } from "../../../DescriptionCard";
+import { ConfigIcon } from "../../../Icon/ConfigIcon";
 import ConfigInsightsDetailsModal from "./ConfigInsightsDetailsModal";
-import { ConfigAnalysis } from "../../../api/types/configs";
+import { ConfigAnalysis } from "../../../../api/types/configs";
 
 type Props = {
   configAnalysis: Pick<
