+++ src/components/Insights/cells/ConfigInsightNameCell.tsx
@@ -1,5 +1,5 @@
 import { CellContext } from "@tanstack/react-table";
-import { ConfigAnalysisLink } from "../../Configs/ConfigAnalysisLink/ConfigAnalysisLink";
+import { ConfigAnalysisLink } from "../../Configs/Insights/ConfigAnalysisLink/ConfigAnalysisLink";
 import { ConfigAnalysis } from "../../../api/types/configs";
 
 export default function ConfigInsightNameCell({
